requirements.txt
pytorch==2.2.0
numpy==1.26.4
einops==0.7.0
h5py==3.10.0
joblib==1.3.2 
wandb==0.16.3